[Gold V] Title: 옥상 정원 꾸미기, Time: 0 ms, Memory: 3240 KB -BaekjoonHub

+++ 백준/Gold/6198. 옥상 정원 꾸미기/옥상 정원 꾸미기.cc
@@ -4,7 +4,7 @@
 #pragma GCC optimize("O3,unroll-loops")
 #pragma GCC target("avx,avx2,fma")
 using namespace std;
-using ll = long long;
+using ll = int;
 using pll = pair<ll,ll>;
 using ld = long double;
 using pld = pair<ld,ld>;
@@ -79,7 +79,7 @@ void solve(ll testcase){
     FastScanner fs;
     ll N = fs.nxtLL();
     vl st; st.reserve(N);
-    ll ans = 0;
+    long long ans = 0;
     for (i=0; i<N; i++) {
         ll num = fs.nxtLL();
         while (!st.empty() && st.back()<=num) st.pop_back();
